implementacion avanzada de file descriptors Co-authored-by: ozacagnino <ozacagnino@users.noreply.github.com> Co-authored-by: Facundo Lasserre <flasserre@itba.edu.ar>

+++ Kernel/include/pipe.h
@@ -18,7 +18,7 @@ typedef struct pipe{
 
 int initPipes();
 
-int pipeCreate();
+uint64_t pipeCreate();
 
 void pipeDestroy(uint16_t pipeId);
